update transactions page

+++ app/Http/Controllers/TransactionController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Transaction;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
+use Yajra\Datatables\Datatables;
+use DB;
 
 class TransactionController extends Controller
 {
@@ -16,6 +18,79 @@ class TransactionController extends Controller
     public function __construct()
     {
         $this->middleware('auth');
+        $this->middleware('only_admin_access')->only('destroy');
+    }
+
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function index()
+    {
+        $title = __('Transactions');
+        $total = Transaction::select('id')->get()->count();
+        
+        return view('transactions.index', compact('title', 'total'));
+    }
+
+    public function transactions(Request $request)
+    {
+        if ($request->ajax()) {
+            $data = DB::table('transactions')->latest()->get();
+            return Datatables::of($data)
+                ->addIndexColumn()
+                ->addColumn('created_on', function($row){
+                    $created_on = date('m/d/Y', strtotime($row->created_at));
+                    return $created_on;
+                })
+                ->addColumn('customer', function($row){
+                    $customer = '<a href="' . route('customers.show', $row->customer_id) . '" class="text-primary">' . $row->customer->first_name . ' ' . $row->customer->last_name . '</a>';
+                    return $customer;
+                })
+                ->addColumn('store_credit', function($row){
+                    if ($row->transaction_type == 'Purchase') {
+                        if ($row->store_credit != 0) {
+                          $store_credit = "-$" . $row->store_credit;
+                        } else {
+                          $store_credit = "$0.00";
+                        }
+                    } else if ($row->transaction_type == 'Cash out for store credit') {
+                        $store_credit = "-$" . $row->cash_out_for_storecredit;
+                    } else {
+                        $store_credit = "$" . $row->store_credit;
+                    }
+                    
+                    return $store_credit;
+                })
+                ->addColumn('cash', function($row){
+                    $cash = number_format($row->cash_in + $row->cash_out_for_trade + $row->cash_out_for_storecredit/2, 2, '.', '');
+                    if ($row->transaction_type == "Cash out for store credit" || $row->transaction_type == "Cash out for trade"){
+                        $cash = "-$" .$cash;
+                    } else {
+                        $cash = "$" .$cash;
+                    }
+                    
+                    return $cash;
+                })
+                ->addColumn('action', function($row){
+                    $actions = 
+                        '<a href="' . route('transactions.show', $row->id) . '" class="btn btn-primary p-2" rel="tooltip" data-original-title="" title="View"><i class="material-icons">visibility</i></a>
+                        ';
+
+                    if(auth()->user()->user_type == 'admin') {
+                        $actions = $actions . '<a href="' . route('transactions.edit', $row->id) . '" class="btn btn-warning p-2" rel="tooltip" data-original-title="" title="Edit"><i class="material-icons">edit</i></a>
+                        <form action="' . route('transactions.destroy',$row->id) . '" method="POST">
+                        <input type="hidden" name="_token" value="' . csrf_token() . '">
+                        <input type="hidden" name="_method" value="delete">
+                        <button type="submit" class="btn btn-danger p-2" onclick="return confirm(Are you sure you want to permanently delete Transaction #' . $row->id . '?\')" rel="tooltip" data-original-title="" title="Delete"><i class="material-icons">delete</i></button>
+                        </form>';
+                    }
+                    return $actions;
+                })
+                ->rawColumns(['phone_number','action'])
+                ->make(true);
+        }
     }
 
     /**
@@ -120,6 +195,9 @@ class TransactionController extends Controller
      */
     public function edit($id)
     {
+        if(auth()->user()->user_type != 'admin') {
+            return redirect()->back();
+        }
         $title = __("Edit Transaction");
         $transaction = Transaction::find($id);
         return view('transactions.edit', compact('title', 'transaction'));
@@ -169,6 +247,9 @@ class TransactionController extends Controller
      */
     public function destroy($id)
     {
+        if(auth()->user()->user_type != 'admin') {
+            return redirect()->back();
+        }
         $transaction = Transaction::find($id);
         $transaction->delete();
         return redirect()->back()->with('success', 'Transaction #' . $id . ' has successfully been deleted.');
